refactor: update utilities and token handling
